requirements.txt
torch
torchvision
streamlit
pillow
opencv-python
